+++ db/migrate/20230511153023_make_contacts_single_table_inheritance.rb
@@ -1,0 +1,16 @@
+class MakeContactsSingleTableInheritance < ActiveRecord::Migration[7.0]
+  def up
+    add_column :contacts, :type, :string
+
+    Contact::Project.all.each do |contact|
+      if contact.project
+        contact.type = "Contact::Project"
+        contact.save(validate: false)
+      end
+    end
+  end
+
+  def down
+    remove_column :contacts, :type, :string
+  end
+end